+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add order payment information in the dashboard, including a specific
   interface displayed on failed installment in order to retry the payment.
 - Display Payment Schedule into SaleTunnel
+- Add OrderHelper allowEnrollment
 
 ### Changed